[FIX] No-transform elements. Ignore elements not in schema.
For some reason, for elements that already are in WCH API format,
"elementValue instanceof Object" was not true. Since we also need
an elementType property inside, changed that test.

Also, if the input had extraneous elements, WCH create API was
happy to take them in, but that causes publishing to fail later.
A better handling is removing them before creating the content
item.